pyproject.toml
[project]
name = "cs6440-practicum-g095"
version = "0.1.0"
description = "AI-powered summarizer for clinical reports using FHIR data"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.24.0",
    "streamlit==1.29.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0,<2.0.0",
    "requests==2.31.0",
    "python-dotenv==1.0.1",
    "matplotlib==3.8.0",
    "plotly==6.1.0",
    "strands-agents>=1.0.1",
    "strands-agents[gemini]",
    "fastmcp>=2.2.0",
    "sentence-transformers>=2.2.2",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
